requirements.txt
allure-behave
allure-pytest
allure-python-commons
anyio
async-generator
attrs
behave
certifi
cffi
colorama
exceptiongroup
execnet
h11
idna
iniconfig
numpy
outcome
packaging
parse
parse-type
pluggy
pycountry
pycparser
PySocks
pytest
pytest-xdist
selenium
six
sniffio
sortedcontainers
tomli
trio
trio-websocket
urllib3
wsproto
junit-xml
